Add files via upload
couette_wetnode_first_order is the same scheme as the original script last week, but with the expressions generalised to arbitrary speed of sound, with the bounce-back conditions generalised, and imposing the continuity equation. couette_wetnode_zou-he is the main action: it includes second-order (Zou-He/nonequilibrium bounce back) conditions.

The results of the original file, first_order, and zou-he are the same for Couette, but Zou-He should distinguish itself in Poiseuille flow.